move the pubsub in the xpub-client directory
